Restore `setindex!(::IRCode, ::SSAValue, ::SSAValue)`
This particular method got removed in f67371dc, but it's common
in RAUW-like patterns in external code. I think it is a sensible
method to have, because the metadata on such an instruction will
get deleted on the next compact anyway, so there's not much reason
to attempt to force the API user to pass complete metadata around.
It also enables useful patterns like:
```
ir[idx] = insert_node!(idx-2, NewInstruction(...))
```